refactor: Abstract note card to component, add more structure to components folder
