+++ src/pages/notes.tsx
@@ -1,8 +1,9 @@
 import { Button, Paper, Title, Text, ActionIcon } from "@mantine/core";
 import { IconPencilPlus, IconEdit, IconTrash } from "@tabler/icons";
 import { type FC, useEffect } from "react";
-import CommonLayout from "../Components/CommonLayout";
-import NotesListLayout from "../Components/NotesListLayout";
+import CommonLayout from "../Components/Layouts/CommonLayout";
+import NotesListLayout from "../Components/Layouts/NotesListLayout";
+import NoteCard from "../Components/NoteCard";
 import { useGlobalModal } from "../hooks/useGlobalModal";
 import { useNotes } from "../hooks/useNotes";
 import { getAllNotesFromLocalStorage } from "../services/notesLocalStorage";
@@ -10,8 +11,8 @@ import { ModalTypes } from "../types/modals";
 import { areObjectEqual } from "../utils/comparsion";
 
 const Notes: FC = () => {
-  const { openModal } = useGlobalModal();
   const { notesState, setAllNotes } = useNotes();
+  const { openModal } = useGlobalModal();
 
   useEffect(() => {
     const notesFromLocaleStorage = getAllNotesFromLocalStorage();
@@ -30,34 +31,7 @@ const Notes: FC = () => {
         {notesState.length ? (
           <NotesListLayout>
             {notesState.map(({ note, title, id }) => (
-              <Paper
-                key={id}
-                className="max-h-[300px] overflow-hidden"
-                shadow="lg"
-                p="md"
-                withBorder
-              >
-                <div className="flex justify-between gap-10">
-                  <Title order={3}>{title}</Title>
-                  <div className="flex gap-2">
-                    <ActionIcon aria-label="Edit note">
-                      <IconEdit size={20} stroke={1} />
-                    </ActionIcon>
-                    <ActionIcon
-                      aria-label="Delete note"
-                      onClick={() =>
-                        openModal({
-                          modalType: ModalTypes.DELETE_NOTE,
-                          payload: { deleteNoteId: id },
-                        })
-                      }
-                    >
-                      <IconTrash size={20} stroke={1} />
-                    </ActionIcon>
-                  </div>
-                </div>
-                <Text>{note}</Text>
-              </Paper>
+              <NoteCard note={note} title={title} key={id} id={id} />
             ))}
           </NotesListLayout>
         ) : (
